Fix build issues on AIX 7.1
mmap is redefined to mmap64 when large file support is enabled.

+++ contrib/reset-cnfs.c
@@ -46,7 +46,8 @@ main(int argc, char *argv[])
                  * fuzz.  buf has 512 bytes in it, therefore containing data
                  * for (512 * 8) * 512 bytes of data. */
                 numwr = (st.st_size / (512*8) / sizeof(buf)) + 50;
-                printf("File %s: %lu %u\n", argv[i], st.st_size, numwr);
+                printf("File %s: %lu %u\n", argv[i],
+                       (long unsigned) st.st_size, numwr);
                 for (j = 0; j < numwr; j++) {
                     if (!(j % 100)) {
                         printf("\t%d/%d\n", j, numwr);
